Armada7k8k: allow overriding the base MAC address.
Clearly no one ever ran more than a McBin in an organization,
because they all come up with the same MAC addresses.

This allows *you* to go into the UEFI setup and set a base
address, used to form all the interfaces (by adding the interface
index number to it).

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>